+++ authority-portal-backend/authority-portal-api/src/main/java/de/sovity/authorityportal/api/model/ProvidedConnectorOverviewEntryDto.java
@@ -1,5 +1,6 @@
 package de.sovity.authorityportal.api.model;
 
+import de.sovity.authorityportal.api.model.organization.ConnectorStatusDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -20,4 +21,6 @@ public class ProvidedConnectorOverviewEntryDto {
     private DeploymentEnvironmentDto environment;
     @Schema(description = "Connector Name", requiredMode = Schema.RequiredMode.REQUIRED)
     private String name;
+    @Schema(description = "Connector status", requiredMode = Schema.RequiredMode.REQUIRED)
+    private ConnectorStatusDto status;
 }