+++ CHANGELOG.md
@@ -39,6 +39,7 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 - Changed Connector URL persistence. Frontend URL, Endpoint URL and Management API URL are now independent of each other
 - slide-overs scrolling issue has been fixed
 - Added a connector overview for service providers and the option to unregister provided connectors
+- Connector status is now fetched from Broker and sovity portal and periodically updated
 
 #### Patch